src/main/java/com/fiap/challenge/tastefood/app/adapter/output/persistence/gateway/CustomerGatewayImpl.java
package com.fiap.challenge.tastefood.app.adapter.output.persistence.gateway;

import com.fiap.challenge.tastefood.app.adapter.output.externalapis.CustomerClient;
import com.fiap.challenge.tastefood.core.domain.Customer;
import com.fiap.challenge.tastefood.core.gateways.CustomerGateway;
import lombok.AllArgsConstructor;
import org.springframework.stereotype.Component;

import java.util.Optional;
import java.util.UUID;

@Component
@AllArgsConstructor
public class CustomerGatewayImpl implements CustomerGateway {

    private CustomerClient client;

    public Optional<Customer> findById(UUID id) {
        return client.getCustomer(id);
    }

}
